backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
python-dotenv
pymongo
dnspython
Pillow
azure-storage-blob==12.24.1